+++ docs/tutorial.rst
@@ -1276,11 +1276,11 @@ Usually, callable alias commands will be run in a separate thread so that
 they may be run in the background.  However, some aliases may need to be
 executed on the thread that they were called from. This is mostly useful for
 debuggers and profilers. To make an alias run in the foreground, decorate its
-function with the ``xonsh.proc.unthreadable`` decorator.
+function with the ``xonsh.tools.unthreadable`` decorator.
 
 .. code-block:: python
 
-    from xonsh.proc import unthreadable
+    from xonsh.tools import unthreadable
 
     @unthreadable
     def _mycmd(args, stdin=None):
@@ -1297,12 +1297,12 @@ Thus the callable alias can't be captured without dire consequences (tm).
 To prevent this, you can declare a callable alias uncapturable. This is mostly
 useful for aliases that then open up text editors, pagers, or the like.
 To make an alias uncapturable, decorate its
-function with the ``xonsh.proc.uncapturable`` decorator. This is probably
+function with the ``xonsh.tools.uncapturable`` decorator. This is probably
 best used in conjunction with the ``unthreadable`` decorator.  For example:
 
 .. code-block:: python
 
-    from xonsh.proc import unthreadable, uncapturable
+    from xonsh.tools import unthreadable, uncapturable
 
     @uncapturable
     @unthreadable
