+++ news/move-tools.rst
@@ -1,0 +1,14 @@
+**Added:** None
+
+**Changed:**
+
+* Moved decorators ``unthreadable``, ``foreground``, ``uncapturable`` from
+  ``xonsh.proc`` to ``xonsh.tools``.
+
+**Deprecated:** None
+
+**Removed:** None
+
+**Fixed:** None
+
+**Security:** None
