+++ tests/test_integrations.py
@@ -148,7 +148,7 @@ with open('tttt', 'w') as fp:
 """, '      1       3      24\n' if ON_WINDOWS else " 1  4 16 <stdin>\n", 0),
 # test unthreadable alias (which should trigger a ProcPoxy call)
 ("""
-from xonsh.proc import unthreadable
+from xonsh.tools import unthreadable
 
 @unthreadable
 def _f():
